Finished the basic component
Finished the basic components:
Damage zones deal damage, you can specify what entities they can damage in the editor, as well as if they deal elemental damage.
Health Component has a hitbox and when hit will keep track of it's current and max health, as well as how many stacks of each element it has. If it goes below 0 health it will die and delete the entity it is attached to, triggering on death effects.