Making the game feel....?
This commit was dedicated to messing with the character controller, bauble controller, and camera. The main goal was to create a finalized draft of the player and baubles and get them to a  point where all of their planned base abilities are in game. I started with the throw and while tweaking that I couldn't figure out why but for some reason no matter what changes I made it never felt natural at all. I also tried to edit the bauble follow code because their following was always very unfinished so they never followed in the correct positions, but ofc that also felt off. I realized that the reason they felt so off was because the camera and viewport were off as while the characters themselves felt ok, the small viewport made everything look like it was moving too fast. Increasing the viewport helped a lot, but I also made changes to the camera so that it would shift towards wherever the player's mouse was, that way when they threw baubles they wouldn't just immediately fly off the screen, and so the player could keep better track of wherever the enemies and baubles were. Those changes made the game feel a lot more natural and reasonable to play, so now I'll move on to actually creating/editing the abilities like intended