Adjusted the throw
Now the throw works on some odd speed math that will calculate how quickly the bauble needs to move in order to sync up it's throw movement and throw animation. It feels much better than before, but at this point I need to do the math to create the throw curve rather than relying on a jank animation for it.